requirements.txt
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
